Additions & Changes to UI, Enemy and Player In this commit I added and modified several things.
A modification to UI was to the GameUI to also have another box container in the top
right to hold icons of the upgrades you find around the world.

Another part of the UI is that I created a Pause Menu with full functionality, as the name implies it
will pause the game, you can change the game's volume, resume and quit the game.

For the enemy I added a raycast down so that it doesn't try to go over
gaps and turns around instead. The enemy also doesn't see the player as
a collidable object and goes through it.

The player also got a couple functions added to it to handle damage from
the enemy (when the enemy goes through the player) to avoid constant
damage the player receives invincibillity for a set amount of time

-- Comment: --
When the player dies the game just quits. Nothing specific has been
added for that.